refactor: 패키지명 명확화
